+++ test.js
@@ -14,17 +14,17 @@ equal(
 
 equal(
   md().use(anchor, { permalink: true }).render('# H1'),
-  '<h1 id="h1">H1 <a class="header-anchor" href="#h1">¶</a></h1>\n'
+  '<h1 id="h1">H1 <a class="header-anchor" href="#h1" aria-hidden="true">¶</a></h1>\n'
 )
 
 equal(
   md().use(anchor, { permalink: true, permalinkClass: 'test' }).render('# H1'),
-  '<h1 id="h1">H1 <a class="test" href="#h1">¶</a></h1>\n'
+  '<h1 id="h1">H1 <a class="test" href="#h1" aria-hidden="true">¶</a></h1>\n'
 )
 
 equal(
   md().use(anchor, { permalink: true, permalinkSymbol: 'P' }).render('# H1'),
-  '<h1 id="h1">H1 <a class="header-anchor" href="#h1">P</a></h1>\n'
+  '<h1 id="h1">H1 <a class="header-anchor" href="#h1" aria-hidden="true">P</a></h1>\n'
 )
 
 equal(
@@ -34,5 +34,5 @@ equal(
 
 equal(
   md().use(anchor, { permalink: true, permalinkBefore: true }).render('# H1'),
-  '<h1 id="h1"><a class="header-anchor" href="#h1">¶</a> H1</h1>\n'
+  '<h1 id="h1"><a class="header-anchor" href="#h1" aria-hidden="true">¶</a> H1</h1>\n'
 )